image_processor_func/requirements.txt
functions-framework==3.8.2
requests>=2.32.3
Pillow==11.1.0
pillow-avif-plugin==1.5.2
cloudevents==1.11.0
packaging==25.0